refactor: 에러 코드 라벨 매핑 공통 유틸 추출
