+++ src/pages/TempWrite/TempWritePage.tsx
@@ -30,12 +30,16 @@ import {
   getSummaryStatus,
   cancelSummary,
   editPost,
-  getTagsByKeyword,
   getPostDetail,
 } from "@/api/post.api";
 import { uploadImage } from "@/api/image.api";
 import { isAxiosError } from "axios";
 import { startRefresh } from "@/api/axios";
+import { canonicalizeTags } from "@/shared/utils/canonicalizeTags";
+import {
+  ERROR_OPTIONS,
+  toErrorLabel,
+} from "@/shared/utils/errorCodeLabel";
 
 // ---- 숫자 인덱스 변환 유틸 ----
 const QI = { ERROR: 0, REASON: 1 } as const;
@@ -107,22 +111,6 @@ type IncomingTemplateState = {
   postId?: number; // 수정 식별용
   mode?: "edit" | "create";
 };
-
-// ---------- 에러 라벨/코드  ----------
-const ERROR_CODE_TO_LABEL: Record<string, string> = {
-  BUILD_COMPILE_ERROR: "Build/Compile Error",
-  RUNTIME_ERROR: "Runtime Error",
-  DEPENDENCY_VERSION_ERROR: "Dependency/Version Error",
-  NETWORK_API_ERROR: "Network/API Error",
-  AUTHENTICATION_AUTHORIZATION_ERROR: "Authentication/Authorization Error",
-  DATABASE_ERROR: "Database Error",
-  UI_RENDERING_ERROR: "UI/Rendering Error",
-  CONFIGURATION_ERROR: "Configuration Error",
-  TIMEOUT_ERROR_HANDLING: "Timeout/Error Handling",
-  THIRD_PARTY_LIBRARY_ERROR: "Third-Party Library Error",
-};
-const toErrorLabel = (code?: string | null) =>
-  code ? ERROR_CODE_TO_LABEL[code] ?? code : null;
 
 // ---------- 체크리스트 ----------
 const enrichBlocksWithChecklist = (
@@ -378,35 +366,6 @@ const TempWritePage = () => {
             summaryType: "NONE",
           } as any)
       );
-
-  // ---------- 태그 정규화 ----------
-  const canonicalizeTags = async (rawTags: string[]) => {
-    const out: string[] = [];
-    const seen = new Set<string>();
-    for (const raw of rawTags) {
-      const q = String(raw).replace(/^#\s*/, "").trim();
-      if (!q) continue;
-
-      const res: any = await getTagsByKeyword({ tagName: q });
-      const list: any[] = Array.isArray(res)
-        ? res
-        : res?.data ?? res?.content ?? res?.results ?? [];
-      const names = list
-        .map((t) => (typeof t === "string" ? t : t?.name ?? t))
-        .filter(Boolean);
-      const exact = names.find(
-        (n: string) => n.toLowerCase() === q.toLowerCase()
-      );
-      const pick = (exact ?? names[0]) as string | undefined;
-
-      const normalized = String(pick ?? q).trim();
-      if (!seen.has(normalized.toLowerCase())) {
-        seen.add(normalized.toLowerCase());
-        out.push(normalized);
-      }
-    }
-    return out;
-  };
 
   // 현재 화면 상태 + PostSaveModal에서 받은 meta로 수정 페이로드 만들기
   const buildEditFormFromMeta = async (
@@ -1068,18 +1027,7 @@ const TempWritePage = () => {
                 />
 
                 <DropDownButton
-                  options={[
-                    "Build/Compile Error",
-                    "Runtime Error",
-                    "Dependency/Version Error",
-                    "Network/API Error",
-                    "Authentication/Authorization Error",
-                    "Database Error",
-                    "UI/Rendering Error",
-                    "Configuration Error",
-                    "Timeout/Error Handling",
-                    "Third-Party Library Error",
-                  ]}
+                  options={ERROR_OPTIONS}
                   placeholder={selectedErrorType ?? "에러 종류를 선택하세요"}
                   width="w-full sm:w-[180px] lg:w-[220px]"
                   onSelect={(v: string) => setSelectedErrorType(v)}
